feat(browser): Extendable actions (#22)
* feat(browser): Extendable actions

You can now add/remove/modify actions that the pool will take on data coming from the worker. It by
default will handle event, property, and release actions.

#21

* feat(webworker): Make websocketworker actions easier to define

Actions in the websocketworker are now in an actions object which will increase access time when
there is lots of chatter. It also makes it easier to add more actions.

#21

* refactor(webworker): Cleanup